Firebase Completed Authentication
Has full firebase integration with a test page. The authentication is completed, user's email and password are saved to the authentication tab in the Firebase project. Log out button also sends the user back to the log in page. The Cloud Firestore Database is the next WIP from this starter implementation.