ViewOtherProfilePage
Essentially the same as the general profile page. Except it has a floating point button that serves as the following button. I tried to make it similar to the one from the figma, but the figma one looked kinda trash. So this time I tried to add a unique design element. It pops up from the app bar kinda like a bookmarked movie ticket