Navigation and File Aggregation
This push has a lot of changes. It aggregates files from all of the other branches to start integrating large portions of the app. Most of the pages are able to be navigated to and many of them have functionality changes such as changing text and icons to buttons to allow for proper navigation. There is now a class for the navigation bar, so use it whenever possible to keep the nav bar on the pages uniform. Auth for the login page doesn't have input validation yet but the final design is implemented.