Avatar Selection & Achievements
Created the Avatar Selection page, the only needed changes are to get the correct images in the assets folder. Started the Achievements logic, right now the function for following achievements works (as in calling on it will check if the user has earned a new followingCount achievement). Todo: implement the other type of achievements logic and figure out how the flash messages / notifications will look like.